+++ persistence/context.go
@@ -109,7 +109,7 @@ func (p *PostgresContext) Release() {
 }
 
 // INVESTIGATE(#361): Revisit if is used correctly in the context of the lifecycle of a persistenceContext and a utilityUnitOfWork
-func (p *PostgresContext) IndexTransaction(txResult coreTypes.TxResult) error {
+func (p *PostgresContext) IndexTransaction(txResult *coreTypes.TxResult) error {
 	return p.txIndexer.Index(txResult)
 }