+++ persistence/test/state_test.go
@@ -7,7 +7,6 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/pokt-network/pocket/persistence/indexer"
 	"github.com/pokt-network/pocket/shared/codec"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/utils"
@@ -28,7 +27,7 @@ type TestReplayableOperation struct {
 }
 type TestReplayableTransaction struct {
 	operations []*TestReplayableOperation
-	txResult   coreTypes.TxResult
+	txResult   *coreTypes.TxResult
 }
 
 type TestReplayableBlock struct {
@@ -73,7 +72,7 @@ func TestStateHash_DeterministicStateWhenUpdatingAppStake(t *testing.T) {
 		require.NoError(t, err)
 
 		txBz := []byte("a tx, i am, which set the app stake amount to " + stakeAmountStr)
-		txResult := indexer.TxRes{
+		txResult := &coreTypes.TxResult{
 			Tx:            txBz,
 			Height:        height,
 			Index:         0,
@@ -84,7 +83,7 @@ func TestStateHash_DeterministicStateWhenUpdatingAppStake(t *testing.T) {
 			MessageType:   "TODO",
 		}
 
-		err = db.IndexTransaction(coreTypes.TxResult(&txResult))
+		err = db.IndexTransaction(txResult)
 		require.NoError(t, err)
 
 		// Update the state hash
@@ -157,7 +156,7 @@ func TestStateHash_ReplayingRandomTransactionsIsDeterministic(t *testing.T) {
 						}
 					}
 
-					txResult := coreTypes.TxResult(getRandomTxResult(height))
+					txResult := getRandomTxResult(height)
 					err := db.IndexTransaction(txResult)
 					require.NoError(t, err)