+++ utility/types/tx_result.go
@@ -3,12 +3,8 @@ package types
 import (
 	"encoding/hex"
 
-	"github.com/pokt-network/pocket/shared/codec"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
-	"github.com/pokt-network/pocket/shared/crypto"
 )
-
-var _ coreTypes.TxResult = &TxResult{}
 
 func TxToTxResult(
 	tx *coreTypes.Transaction,
@@ -16,7 +12,7 @@ func TxToTxResult(
 	index int,
 	msg Message,
 	msgHandlingResult Error,
-) (*TxResult, Error) {
+) (*coreTypes.TxResult, Error) {
 	txBz, err := tx.Bytes()
 	if err != nil {
 		return nil, ErrProtoMarshal(err)
@@ -27,7 +23,7 @@ func TxToTxResult(
 		resultCode = int32(msgHandlingResult.Code())
 		errorMsg = msgHandlingResult.Error()
 	}
-	return &TxResult{
+	result := &coreTypes.TxResult{
 		Tx:            txBz,
 		Height:        height,
 		Index:         int32(index),
@@ -36,29 +32,6 @@ func TxToTxResult(
 		SignerAddr:    hex.EncodeToString(msg.GetSigner()),
 		RecipientAddr: msg.GetMessageRecipient(),
 		MessageType:   msg.GetMessageName(),
-	}, nil
-}
-
-func (txr *TxResult) Bytes() ([]byte, error) {
-	return codec.GetCodec().Marshal(txr)
-}
-
-func (*TxResult) FromBytes(bz []byte) (coreTypes.TxResult, error) {
-	result := new(TxResult)
-	if err := codec.GetCodec().Unmarshal(bz, result); err != nil {
-		return nil, err
 	}
 	return result, nil
-}
-
-func (txr *TxResult) Hash() ([]byte, error) {
-	bz, err := txr.Bytes()
-	if err != nil {
-		return nil, err
-	}
-	return txr.HashFromBytes(bz)
-}
-
-func (txr *TxResult) HashFromBytes(bz []byte) ([]byte, error) {
-	return crypto.SHA3Hash(bz), nil
 }