+++ utility/unit_of_work/transaction.go
@@ -15,7 +15,7 @@ import (
 //
 // IMPROVE: hydration should accept and return the same type (i.e. TxResult) so there may be opportunity
 // to refactor this in the future.
-func (u *baseUtilityUnitOfWork) hydrateTxResult(tx *coreTypes.Transaction, index int) (coreTypes.TxResult, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) hydrateTxResult(tx *coreTypes.Transaction, index int) (*coreTypes.TxResult, typesUtil.Error) {
 	msg, err := u.anteHandleMessage(tx)
 	if err != nil {
 		return nil, err