+++ utility/doc/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.37] - 2023-04-13
+
+- Remove utility specific `TxResult` protobuf and interface
+- Utlise the `TxResult` protobuf in `shared/core/types`
+
 ## [0.0.0.36] - 2023-04-07
 
 - Add wrapper functions for `GetParameter()`